Update AzSPoC to fix a failure in 2102 during the RegisterNewRPs step (#139)
Starting in 2102 the Metering and Consumption subscriptions are disabled, and if you try to register RPs in a disabled subscription you will fail with the following error: "The subscription '...' is disabled and therefore marked as read only. You cannot perform any write actions on this subscription until it is re-enabled."  By adding a where to filter out any disabled subscriptions we avoid this error.